test: split test_listeners.py by test class
tests/integration/test_listeners.py was 832 lines, over the file-sizes
CI check's 800-line threshold (HSL102). Split its 14 independent test
classes into three themed files, moving the shared MockEvent helper to a
sibling module.

- test_listeners_dispatch.py: invoke, integration, dispatch/cancel
- test_listeners_timing.py: debounce, throttle, rate-limiter cancel
- test_listeners_validation.py: dependency errors, option validation,
  registration metadata
- listener_helpers.py: MockEvent and mock_event()

Test bodies are moved verbatim; no assertions changed. Same 49 tests
collected before and after.

Closes #1755
