+++ tests/integration/listener_helpers.py
@@ -1,0 +1,20 @@
+"""Shared helpers for listener integration tests (test_listeners_*.py)."""
+
+from dataclasses import dataclass
+
+from hassette.events import Event
+
+
+@dataclass(frozen=True, slots=True)
+class MockEvent(Event[str]):
+    """Mock event for testing."""
+
+    @property
+    def data(self) -> str:
+        """Return payload for backward compatibility with tests."""
+        return self.payload
+
+
+def mock_event(data: str = "test") -> MockEvent:
+    """Create a MockEvent with a topic and payload."""
+    return MockEvent(topic="test_topic", payload=data)
