--- tests/integration/test_listeners.py
@@ -1,832 +1,0 @@
-import asyncio
-from dataclasses import dataclass
-
-import pytest
-
-from hassette import D
-from hassette.bus.listeners import ListenerOptions
-from hassette.bus.rate_limiter import RateLimiter
-from hassette.events import Event
-from hassette.exceptions import DependencyResolutionError
-from hassette.models import states
-from hassette.task_bucket import TaskBucket
-from hassette.test_utils import (
-    make_controlled_clock,
-    make_full_state_change_event,
-    make_light_state_dict,
-    make_state_dict,
-    wait_for,
-)
-from hassette.test_utils.helpers import create_listener, settle
-
-
-@dataclass(frozen=True, slots=True)
-class MockEvent(Event[str]):
-    """Mock event for testing."""
-
-    @property
-    def data(self) -> str:
-        """Return payload for backward compatibility with tests."""
-        return self.payload
-
-
-def mock_event(data: str = "test") -> MockEvent:
-    """Create a MockEvent with a topic and payload."""
-    return MockEvent(topic="test_topic", payload=data)
-
-
-class TestListenerInvoke:
-    """Test Listener.invoke() — dependency injection and handler invocation."""
-
-    async def test_sync_handler_with_event(self, bucket: TaskBucket):
-        """Test sync handler that expects an event."""
-        calls = []
-
-        def handler(event: MockEvent):
-            calls.append(event.data)
-
-        listener = create_listener(handler, task_bucket=bucket, owner_id="test", topic="t")
-        await listener.invoker.invoke(mock_event("test_data"))
-        assert calls == ["test_data"], "Handler should be called with event data"
-
-    async def test_async_handler_with_event(self, bucket: TaskBucket):
-        """Test async handler that expects an event."""
-        calls = []
-
-        async def handler(event: MockEvent):
-            calls.append(event.data)
-
-        listener = create_listener(handler, task_bucket=bucket, owner_id="test", topic="t")
-        await listener.invoker.invoke(mock_event("test_data"))
-        assert calls == ["test_data"], "Handler should be called with event data"
-
-    async def test_sync_handler_no_event(self, bucket: TaskBucket):
-        """Test sync handler that doesn't expect an event."""
-        calls = []
-
-        def handler():
-            calls.append("called")
-
-        listener = create_listener(handler, task_bucket=bucket, owner_id="test", topic="t")
-        await listener.invoker.invoke(mock_event("test_data"))
-        assert calls == ["called"], "Handler should be called without event data"
-
-    async def test_async_handler_no_event(self, bucket: TaskBucket):
-        """Test async handler that doesn't expect an event."""
-        calls = []
-
-        async def handler():
-            calls.append("called")
-
-        listener = create_listener(handler, task_bucket=bucket, owner_id="test", topic="t")
-        await listener.invoker.invoke(mock_event("test_data"))
-        assert calls == ["called"], "Handler should be called without event data"
-
-
-class TestDebounceLogic:
-    """Test debounce functionality via RateLimiter directly."""
-
-    async def test_debounce_delays_execution(self, bucket: TaskBucket):
-        """Test that debounce delays execution until quiet period."""
-        calls: list[str] = []
-
-        def make_handler(label: str):
-            async def handler():
-                calls.append(label)
-
-            return handler
-
-        limiter = RateLimiter(bucket, debounce=0.1)
-
-        await limiter.call(make_handler("first"))
-        await limiter.call(make_handler("second"))
-        await limiter.call(make_handler("third"))
-
-        await asyncio.sleep(0)
-        assert calls == [], "No calls should be made immediately due to debounce"
-
-        await wait_for(lambda: calls == ["third"], desc="debounce fired")
-
-    async def test_debounce_with_no_args(self, bucket: TaskBucket):
-        """Test debounce with a no-arg handler."""
-        calls: list[str] = []
-
-        async def handler():
-            calls.append("called")
-
-        limiter = RateLimiter(bucket, debounce=0.1)
-
-        await limiter.call(handler)
-        await limiter.call(handler)
-        await limiter.call(handler)
-
-        await wait_for(lambda: calls == ["called"], desc="debounce fired")
-
-    async def test_debounce_cancels_previous_calls(self, bucket: TaskBucket):
-        """Test that new debounce calls cancel previous pending calls."""
-        calls: list[str] = []
-
-        def make_handler(label: str):
-            async def handler():
-                calls.append(label)
-
-            return handler
-
-        limiter = RateLimiter(bucket, debounce=0.2)
-
-        await limiter.call(make_handler("first"))
-        # timing: mid-debounce assertion — must be within the 0.2s window
-        await asyncio.sleep(0.1)
-        assert limiter._debounce_task is not None, "Debounce task should be created"
-        assert not limiter._debounce_task.done(), "Debounce task should still be pending"
-
-        await limiter.call(make_handler("second"))
-        # timing: mid-debounce assertion — must be within the 0.2s window
-        await asyncio.sleep(0.1)
-        assert not limiter._debounce_task.done(), "Debounce task should still be pending"
-
-        await limiter.call(make_handler("third"))
-
-        await wait_for(lambda: calls == ["third"], desc="debounce fired")
-        # done_callback is scheduled via call_soon, so it may not have run yet
-        await wait_for(lambda: limiter._debounce_task is None, desc="debounce task reference cleared")
-
-    async def test_debounce_handler_cancelled_error_propagates(self, bucket: TaskBucket):
-        """CancelledError during handler execution must propagate (not be suppressed).
-
-        Debounce reset (cancel during sleep) should be silent, but handler cancellation
-        (e.g., shutdown) should propagate so telemetry can record it as 'cancelled'.
-        """
-
-        async def handler_that_gets_cancelled():
-            raise asyncio.CancelledError()
-
-        limiter = RateLimiter(bucket, debounce=0.01)
-        await limiter.call(handler_that_gets_cancelled)
-
-        # Capture task reference before done_callback clears it
-        task = limiter._debounce_task
-        assert task is not None
-
-        await wait_for(lambda: task.done(), desc="debounce task completed")
-
-        # The task should show as cancelled (CancelledError propagated out of delayed_call)
-        assert task.done()
-        assert task.cancelled(), "Handler CancelledError should propagate, not be suppressed"
-
-    async def test_debounce_reset_cancellation_is_silent(self, bucket: TaskBucket):
-        """CancelledError from debounce reset (new event superseding old) should be silent."""
-        calls: list[str] = []
-
-        def make_handler(label: str):
-            async def handler():
-                calls.append(label)
-
-            return handler
-
-        limiter = RateLimiter(bucket, debounce=0.1)
-
-        # First call starts debounce
-        await limiter.call(make_handler("first"))
-        first_task = limiter._debounce_task
-        assert first_task is not None
-
-        # Second call cancels first (debounce reset)
-        await limiter.call(make_handler("second"))
-        await asyncio.sleep(0)  # Let cancellation propagate
-
-        # First task should be cancelled silently (no crash)
-        assert first_task.cancelled() or first_task.done()
-
-        await wait_for(lambda: calls == ["second"], desc="debounce fired")
-
-
-class TestRateLimiterCancel:
-    """Test RateLimiter.cancel() for cleanup on listener removal."""
-
-    async def test_cancel_pending_debounce(self, bucket: TaskBucket):
-        """Cancelling a pending debounce prevents the handler from firing."""
-        calls: list[str] = []
-
-        async def handler():
-            calls.append("fired")
-
-        limiter = RateLimiter(bucket, debounce=0.5)
-        await limiter.call(handler)
-        assert limiter._debounce_task is not None
-
-        limiter.cancel()
-        assert limiter._debounce_task is None
-
-        await settle(0.6)
-        assert calls == [], "Handler should not fire after cancel"
-
-    async def test_cancel_when_no_task(self, bucket: TaskBucket):
-        """Cancelling with no pending task should not raise."""
-        limiter = RateLimiter(bucket, debounce=0.1)
-        limiter.cancel()  # Should not raise
-
-    async def test_cancel_after_task_completed(self, bucket: TaskBucket):
-        """Cancelling after the task has already completed should not raise."""
-        calls: list[str] = []
-
-        async def handler():
-            calls.append("fired")
-
-        limiter = RateLimiter(bucket, debounce=0.01)
-        await limiter.call(handler)
-        await wait_for(lambda: calls == ["fired"], desc="debounce fired")
-
-        limiter.cancel()  # Should not raise; task already done
-
-
-class TestThrottleLogic:
-    """Test throttle functionality via RateLimiter directly."""
-
-    async def test_throttle_limits_execution_frequency(self, bucket: TaskBucket):
-        """Test that throttle limits how often handler is called."""
-        calls: list[str] = []
-
-        def make_handler(label: str):
-            async def handler():
-                calls.append(label)
-
-            return handler
-
-        clock = make_controlled_clock()
-        limiter = RateLimiter(bucket, throttle=0.1, clock=clock)
-
-        await limiter.call(make_handler("first"))
-        assert calls == ["first"], "First call should be executed immediately"
-
-        await limiter.call(make_handler("second"))
-        await limiter.call(make_handler("third"))
-        assert calls == ["first"], "Subsequent calls should be ignored"
-
-        # advance the controlled clock past the throttle window before the next call
-        clock.advance_to(1.2)
-
-        await limiter.call(make_handler("fourth"))
-        assert calls == ["first", "fourth"], "Fourth call should execute after throttle period"
-
-    async def test_throttle_allows_first_call_with_zero_origin_clock(self, bucket: TaskBucket):
-        """A clock that legitimately returns 0.0 on its first call must not drop that call.
-
-        Regression test: `_throttle_last_time` used to default to `0.0`, so `now - 0.0 <
-        throttle` was true on the very first call whenever `now` was also `0.0`, silently
-        dropping it. `_throttle_last_time` now starts as `None` and the elapsed-window
-        check is bypassed until a timestamp has actually been recorded.
-        """
-        calls: list[str] = []
-
-        def make_handler(label: str):
-            async def handler():
-                calls.append(label)
-
-            return handler
-
-        clock = make_controlled_clock(start=0.0)
-        limiter = RateLimiter(bucket, throttle=0.1, clock=clock)
-
-        await limiter.call(make_handler("first"))
-        assert calls == ["first"], "First call must execute even when the clock starts at 0.0"
-
-    async def test_throttle_with_no_args(self, bucket: TaskBucket):
-        """Test throttle with a no-arg handler."""
-        calls: list[str] = []
-        label = "called"
-
-        async def handler():
-            calls.append(label)
-
-        clock = make_controlled_clock()
-        limiter = RateLimiter(bucket, throttle=0.1, clock=clock)
-
-        await limiter.call(handler)
-        assert calls == ["called"]
-
-        label = "called while throttled"
-        await limiter.call(handler)
-        await limiter.call(handler)
-        assert calls == ["called"]
-
-        label = "called after throttle"
-        # advance the controlled clock past the throttle window before the next call
-        clock.advance_to(1.2)
-        await limiter.call(handler)
-        assert calls == ["called", "called after throttle"]
-
-    async def test_throttle_tracks_time_correctly(self, bucket: TaskBucket):
-        """Test that throttle timing works correctly using an injected clock."""
-        calls: list[str] = []
-
-        def make_handler(label: str):
-            async def handler():
-                calls.append(label)
-
-            return handler
-
-        clock = make_controlled_clock(start=1000.0)
-
-        limiter = RateLimiter(bucket, throttle=0.05, clock=clock)
-
-        await limiter.call(make_handler("1"))
-        assert calls == ["1"]
-
-        clock.advance_to(1000.03)
-        await limiter.call(make_handler("2"))
-        assert calls == ["1"]
-
-        clock.advance_to(1000.06)
-        await limiter.call(make_handler("3"))
-        assert calls == ["1", "3"]
-
-    async def test_throttle_does_not_block_during_handler(self, bucket: TaskBucket):
-        """A second throttled call within the window must not block on the first handler."""
-        handler_started = asyncio.Event()
-        handler_release = asyncio.Event()
-
-        async def slow_handler():
-            handler_started.set()
-            await handler_release.wait()
-
-        limiter = RateLimiter(bucket, throttle=5.0)
-
-        task1 = asyncio.create_task(limiter.call(slow_handler))
-        await handler_started.wait()
-
-        task2 = asyncio.create_task(limiter.call(slow_handler))
-        done, _ = await asyncio.wait({task2}, timeout=0.05)
-        assert task2 in done, "Throttled call within window must return immediately, not block"
-
-        handler_release.set()
-        await task1
-
-
-class TestListenerIntegration:
-    """Test Listener integration with rate limiting."""
-
-    async def test_listener_with_debounce(self, bucket: TaskBucket):
-        """Test Listener with debounce via rate limiter (as BusService._dispatch would)."""
-        calls: list[str] = []
-
-        def handler(event: MockEvent):
-            calls.append(event.data)
-
-        listener = create_listener(
-            handler,
-            task_bucket=bucket,
-            owner_id="test",
-            topic="test_topic",
-            debounce=0.1,
-        )
-
-        assert listener.invoker.rate_limiter is not None
-        rl = listener.invoker.rate_limiter
-
-        # Simulate dispatch: rate_limiter.call(invoke_fn) — like _dispatch does
-        async def invoke_fn():
-            await listener.invoker.invoke(mock_event("1"))
-
-        async def invoke_fn2():
-            await listener.invoker.invoke(mock_event("2"))
-
-        async def invoke_fn3():
-            await listener.invoker.invoke(mock_event("3"))
-
-        await rl.call(invoke_fn)
-        await rl.call(invoke_fn2)
-        await rl.call(invoke_fn3)
-
-        await wait_for(lambda: calls == ["3"], desc="debounce fired")
-
-    async def test_listener_with_throttle(self, bucket: TaskBucket):
-        """Test Listener with throttle via rate limiter (as BusService._dispatch would)."""
-        calls: list[str] = []
-
-        def handler(event: MockEvent):
-            calls.append(event.data)
-
-        clock = make_controlled_clock()
-
-        listener = create_listener(
-            handler,
-            task_bucket=bucket,
-            owner_id="test",
-            topic="test_topic",
-            throttle=0.1,
-            clock=clock,
-        )
-
-        assert listener.invoker.rate_limiter is not None
-        rl = listener.invoker.rate_limiter
-
-        events = [mock_event("1"), mock_event("2"), mock_event("3"), mock_event("4")]
-
-        def make_invoke(ev):
-            async def invoke_fn():
-                await listener.invoker.invoke(ev)
-
-            return invoke_fn
-
-        await rl.call(make_invoke(events[0]))
-        await rl.call(make_invoke(events[1]))
-        await rl.call(make_invoke(events[2]))
-        assert calls == ["1"], "First call should be executed immediately"
-
-        # advance the controlled clock past the throttle window before the next call
-        clock.advance_to(1.2)
-        await rl.call(make_invoke(events[3]))
-        assert calls == ["1", "4"], "Second call should execute after throttle period"
-
-    async def test_listener_without_rate_limiting(self, bucket: TaskBucket):
-        """Test Listener without debounce or throttle."""
-        calls = []
-
-        def handler(event: MockEvent):
-            calls.append(event.data)
-
-        listener = create_listener(handler, task_bucket=bucket, owner_id="test", topic="test_topic")
-
-        # All calls should execute immediately
-        await listener.invoker.invoke(mock_event("1"))
-        await listener.invoker.invoke(mock_event("2"))
-        await listener.invoker.invoke(mock_event("3"))
-
-        assert calls == ["1", "2", "3"], "All calls should be executed immediately"
-
-    async def test_cannot_specify_both_debounce_and_throttle(self, bucket: TaskBucket):
-        """Test that specifying both debounce and throttle raises an error."""
-
-        def handler(event):
-            pass
-
-        with pytest.raises(ValueError, match="Cannot specify both 'debounce' and 'throttle'"):
-            create_listener(
-                handler,
-                task_bucket=bucket,
-                owner_id="test",
-                topic="test_topic",
-                debounce=0.1,
-                throttle=0.1,
-            )
-
-
-class TestListenerDispatchAndCancel:
-    """Test Listener.dispatch() and Listener.cancel() — the public rate limiting API."""
-
-    async def test_dispatch_without_rate_limiter_calls_invoke_fn_directly(self, bucket: TaskBucket):
-        """dispatch() with no rate limiter calls the invoke function immediately."""
-        calls = []
-
-        def handler(event: MockEvent):
-            calls.append(event.data)
-
-        listener = create_listener(handler, task_bucket=bucket, owner_id="test", topic="t")
-
-        async def invoke_fn():
-            await listener.invoker.invoke(mock_event("direct"))
-
-        await listener.invoker.dispatch(invoke_fn)
-        assert calls == ["direct"]
-
-    async def test_dispatch_with_debounce_coalesces(self, bucket: TaskBucket):
-        """dispatch() with debounce coalesces rapid calls — only last fires."""
-        calls: list[str] = []
-
-        def handler(event: MockEvent):
-            calls.append(event.data)
-
-        listener = create_listener(handler, task_bucket=bucket, owner_id="test", topic="t", debounce=0.1)
-
-        for i in range(3):
-
-            async def invoke_fn(val=str(i + 1)):
-                await listener.invoker.invoke(mock_event(val))
-
-            await listener.invoker.dispatch(invoke_fn)
-
-        await wait_for(lambda: calls == ["3"], desc="debounce fired")
-
-    async def test_dispatch_with_throttle_drops_extras(self, bucket: TaskBucket):
-        """dispatch() with throttle allows first call, drops subsequent within window."""
-        calls: list[str] = []
-
-        def handler(event: MockEvent):
-            calls.append(event.data)
-
-        listener = create_listener(handler, task_bucket=bucket, owner_id="test", topic="t", throttle=5.0)
-
-        for i in range(3):
-
-            async def invoke_fn(val=str(i + 1)):
-                await listener.invoker.invoke(mock_event(val))
-
-            await listener.invoker.dispatch(invoke_fn)
-
-        assert calls == ["1"], "Only the first call should execute"
-
-    async def test_dispatch_once_fires_only_once(self, bucket: TaskBucket):
-        """dispatch() on a once=True listener fires the handler exactly once, even without BusService."""
-        calls: list[str] = []
-
-        def handler(event: MockEvent):
-            calls.append(event.data)
-
-        listener = create_listener(handler, task_bucket=bucket, owner_id="test", topic="t", once=True)
-
-        for i in range(3):
-
-            async def invoke_fn(val=str(i + 1)):
-                await listener.invoker.invoke(mock_event(val))
-
-            await listener.invoker.dispatch(invoke_fn)
-
-        assert calls == ["1"], "Once-listener should fire exactly once via dispatch()"
-
-    async def test_cancel_with_rate_limiter_delegates(self, bucket: TaskBucket):
-        """cancel() delegates to the rate limiter's cancel."""
-        listener = create_listener(
-            lambda _e: None,
-            task_bucket=bucket,
-            owner_id="test",
-            topic="t",
-            debounce=0.5,
-        )
-        assert listener.invoker.rate_limiter is not None
-        assert not listener.invoker.rate_limiter._cancelled
-
-        listener.cancel()
-        assert listener.invoker.rate_limiter._cancelled
-
-    async def test_cancel_without_rate_limiter_is_noop(self, bucket: TaskBucket):
-        """cancel() on a listener without rate limiter does not raise."""
-        listener = create_listener(lambda _e: None, task_bucket=bucket, owner_id="test", topic="t")
-        assert listener.invoker.rate_limiter is None
-        listener.cancel()  # should not raise
-
-    async def test_cancel_is_idempotent(self, bucket: TaskBucket):
-        """Calling cancel() twice does not raise."""
-        listener = create_listener(
-            lambda _e: None,
-            task_bucket=bucket,
-            owner_id="test",
-            topic="t",
-            throttle=1.0,
-        )
-        listener.cancel()
-        listener.cancel()  # second call should not raise
-
-
-class TestDependencyValidationErrors:
-    """Test that listeners properly handle dependency resolution errors."""
-
-    async def test_required_state_with_none_raises_error(self, bucket: TaskBucket):
-        """Test that using StateNew with None value raises DependencyResolutionError."""
-        old_state = make_state_dict(entity_id="test.entity", state="off")
-        state_change_event = make_full_state_change_event("test.entity", old_state, None)
-
-        calls = []
-
-        def handler(new_state: D.StateNew[states.BaseState]):
-            calls.append(new_state)
-
-        listener = create_listener(handler, task_bucket=bucket, owner_id="test", topic="t")
-
-        with pytest.raises(DependencyResolutionError):
-            await listener.invoker.invoke(state_change_event)
-
-        assert len(calls) == 0
-
-    async def test_maybe_state_with_none_succeeds(self, bucket: TaskBucket):
-        """Test that using MaybeStateNew with None value succeeds."""
-        old_state = make_state_dict(entity_id="test.entity", state="off")
-        state_change_event = make_full_state_change_event("test.entity", old_state, None)
-
-        calls = []
-
-        def handler(new_state: D.MaybeStateNew[states.BaseState]):
-            calls.append(new_state)
-
-        listener = create_listener(handler, task_bucket=bucket, owner_id="test", topic="t")
-        await listener.invoker.invoke(state_change_event)
-
-        assert len(calls) == 1
-        assert calls[0] is None
-
-    async def test_mixed_maybe_and_required_all_succeed(self, bucket: TaskBucket):
-        """Test handler with both Maybe and required deps when all resolve."""
-        old_state = make_state_dict(entity_id="test.entity", state="off")
-        new_state = make_state_dict(entity_id="test.entity", state="on")
-        state_change_event = make_full_state_change_event("test.entity", old_state, new_state)
-
-        results = []
-
-        def handler(
-            new_state: D.StateNew[states.BaseState],
-            old_state: D.MaybeStateOld[states.BaseState],
-            entity_id: D.EntityId,
-        ):
-            results.append((new_state, old_state, entity_id))
-
-        listener = create_listener(handler, task_bucket=bucket, owner_id="test", topic="t")
-        await listener.invoker.invoke(state_change_event)
-
-        assert len(results) == 1
-        new, old, eid = results[0]
-        assert new is not None
-        assert old is not None
-        assert eid == "test.entity"
-
-    async def test_multiple_required_deps_first_fails(self, bucket: TaskBucket):
-        """Test that if first required dep fails, handler is not called."""
-        old_dict = make_light_state_dict("light.test", "on", brightness=100)
-        event = make_full_state_change_event("light.test", old_dict, None)
-
-        calls = []
-
-        def handler(new_state: D.StateNew[states.BaseState], entity_id: D.EntityId):
-            calls.append((new_state, entity_id))
-
-        listener = create_listener(handler, task_bucket=bucket, owner_id="test", topic="t")
-
-        with pytest.raises(DependencyResolutionError):
-            await listener.invoker.invoke(event)
-
-        assert len(calls) == 0
-
-
-class TestListenerAppKeyAndInstanceIndex:
-    """Test app_key and instance_index fields on Listener."""
-
-    async def test_listener_has_app_key_and_instance_index(self, bucket: TaskBucket) -> None:
-        """Create a Listener via Listener.create() with explicit app_key and instance_index."""
-
-        def handler(event: MockEvent) -> None:
-            pass
-
-        listener = create_listener(
-            handler,
-            task_bucket=bucket,
-            owner_id="MyApp.MyApp.0",
-            topic="test_topic",
-            app_key="my_app",
-            instance_index=1,
-        )
-
-        assert listener.identity.app_key == "my_app"
-        assert listener.identity.instance_index == 1
-        assert listener.identity.owner_id == "MyApp.MyApp.0"
-
-    async def test_listener_defaults_empty_app_key(self, bucket: TaskBucket) -> None:
-        """Create a Listener without app_key, verify it defaults to empty string."""
-
-        def handler(event: MockEvent) -> None:
-            pass
-
-        listener = create_listener(handler, task_bucket=bucket, owner_id="test", topic="test_topic")
-
-        assert listener.identity.app_key == ""
-        assert listener.identity.instance_index == 0
-
-
-class TestOnceWithRateLimitingProhibited:
-    """once=True combined with debounce or throttle is semantically contradictory and must raise."""
-
-    async def test_once_with_debounce_raises_value_error(self, bucket: TaskBucket):
-        async def handler(event):
-            pass
-
-        with pytest.raises(ValueError, match=r"once.*debounce.*throttle"):
-            create_listener(
-                handler,
-                task_bucket=bucket,
-                owner_id="test",
-                topic="test_topic",
-                once=True,
-                debounce=1.0,
-            )
-
-    async def test_once_with_throttle_raises_value_error(self, bucket: TaskBucket):
-        async def handler(event):
-            pass
-
-        with pytest.raises(ValueError, match=r"once.*debounce.*throttle"):
-            create_listener(
-                handler,
-                task_bucket=bucket,
-                owner_id="test",
-                topic="test_topic",
-                once=True,
-                throttle=1.0,
-            )
-
-    async def test_once_without_rate_limiting_is_allowed(self, bucket: TaskBucket):
-        async def handler(event):
-            pass
-
-        listener = create_listener(
-            handler,
-            task_bucket=bucket,
-            owner_id="test",
-            topic="test_topic",
-            once=True,
-        )
-        assert listener.options.once is True
-
-    async def test_rate_limiting_without_once_is_allowed(self, bucket: TaskBucket):
-        async def handler(event):
-            pass
-
-        listener = create_listener(
-            handler,
-            task_bucket=bucket,
-            owner_id="test",
-            topic="test_topic",
-            debounce=1.0,
-        )
-        assert listener.invoker.rate_limiter is not None
-
-
-class TestRateLimitValueValidation:
-    """debounce and throttle must be positive floats -- zero and negative are rejected."""
-
-    async def test_debounce_zero_raises(self, bucket: TaskBucket):
-        async def handler(event):
-            pass
-
-        with pytest.raises(ValueError, match=r"debounce.*positive"):
-            create_listener(handler, task_bucket=bucket, owner_id="test", topic="t", debounce=0.0)
-
-    async def test_throttle_zero_raises(self, bucket: TaskBucket):
-        async def handler(event):
-            pass
-
-        with pytest.raises(ValueError, match=r"throttle.*positive"):
-            create_listener(handler, task_bucket=bucket, owner_id="test", topic="t", throttle=0.0)
-
-    async def test_debounce_negative_raises(self, bucket: TaskBucket):
-        async def handler(event):
-            pass
-
-        with pytest.raises(ValueError, match=r"debounce.*positive"):
-            create_listener(handler, task_bucket=bucket, owner_id="test", topic="t", debounce=-1.0)
-
-    async def test_throttle_negative_raises(self, bucket: TaskBucket):
-        async def handler(event):
-            pass
-
-        with pytest.raises(ValueError, match=r"throttle.*positive"):
-            create_listener(handler, task_bucket=bucket, owner_id="test", topic="t", throttle=-1.0)
-
-
-class TestMarkRegistered:
-    """Test Listener.mark_registered() — one-time db_id assignment."""
-
-    async def test_mark_registered_sets_db_id(self, bucket: TaskBucket) -> None:
-        """mark_registered() sets db_id on first call."""
-        listener = create_listener(lambda _e: None, task_bucket=bucket, owner_id="test", topic="t")
-        assert listener.db_id is None
-
-        listener.mark_registered(42)
-        assert listener.db_id == 42
-
-    async def test_mark_registered_warns_on_double_call(self, bucket: TaskBucket) -> None:
-        """mark_registered() keeps the original db_id when called a second time."""
-        listener = create_listener(lambda _e: None, task_bucket=bucket, owner_id="test", topic="t")
-        listener.mark_registered(42)
-        listener.mark_registered(99)
-
-        assert listener.db_id == 42
-
-
-class TestMarkFired:
-    """Test Listener.mark_fired() — once-guard flag."""
-
-    async def test_mark_fired_sets_fired(self, bucket: TaskBucket) -> None:
-        """mark_fired() sets the internal _fired flag."""
-        listener = create_listener(lambda _e: None, task_bucket=bucket, owner_id="test", topic="t", once=True)
-        assert listener.invoker.fired is False
-
-        listener.invoker.mark_fired()
-        assert listener.invoker.fired is True
-
-
-class TestValidateOptions:
-    """Test ListenerOptions validation via __post_init__."""
-
-    def test_rejects_negative_debounce(self) -> None:
-        with pytest.raises(ValueError, match=r"debounce.*positive"):
-            ListenerOptions(once=False, debounce=-1.0, throttle=None)
-
-    def test_rejects_both_debounce_and_throttle(self) -> None:
-        with pytest.raises(ValueError, match=r"Cannot specify both"):
-            ListenerOptions(once=False, debounce=1.0, throttle=1.0)
-
-    def test_rejects_once_with_debounce(self) -> None:
-        with pytest.raises(ValueError, match=r"once.*debounce.*throttle"):
-            ListenerOptions(once=True, debounce=1.0, throttle=None)
-
-    def test_accepts_valid_options(self) -> None:
-        ListenerOptions(once=False, debounce=1.0, throttle=None)
-        ListenerOptions(once=True, debounce=None, throttle=None)
